refactor: seperate list api
