+++ frappe/website/doctype/web_form/web_list.py
@@ -1,0 +1,58 @@
+from __future__ import unicode_literals
+import frappe, json, os
+from frappe.website.website_generator import WebsiteGenerator
+from frappe import _, scrub
+from frappe.utils import cstr
+from frappe.website.utils import get_comment_list
+from frappe.custom.doctype.customize_form.customize_form import docfield_properties
+from frappe.core.doctype.file.file import get_max_file_size
+from frappe.core.doctype.file.file import remove_file_by_url
+from frappe.modules.utils import export_module_json, get_doc_module
+from six.moves.urllib.parse import urlencode
+from frappe.integrations.utils import get_payment_gateway_controller
+from six import iteritems
+
+@frappe.whitelist()
+def get_in_list_view_fields(doctype):
+	meta = frappe.get_meta(doctype)
+	fields = []
+
+	if meta.title_field:
+		fields.append(meta.title_field)
+	else:
+		fields.append('name')
+
+	if meta.has_field('status'):
+		fields.append('status')
+
+	fields += [df.fieldname for df in meta.fields if df.in_list_view and df.fieldname not in fields]
+
+	def get_field_df(fieldname):
+		if fieldname == 'name':
+			return { 'label': 'Name', 'fieldname': 'name', 'fieldtype': 'Data' }
+		return meta.get_field(fieldname).as_dict()
+
+	return [get_field_df(f) for f in fields]
+
+
+@frappe.whitelist()
+def delete_multiple(web_form_name, docnames):
+	web_form = frappe.get_doc("Web Form", web_form_name)
+
+	docnames = json.loads(docnames)
+
+	allowed_docnames = []
+	restricted_docnames = []
+
+	for docname in docnames:
+		owner = frappe.db.get_value(web_form.doc_type, docname, "owner")
+		if frappe.session.user == owner and web_form.allow_delete:
+			allowed_docnames.append(docname)
+		else:
+			restricted_docnames.append(docname)
+
+	for docname in allowed_docnames:
+		frappe.delete_doc(web_form.doc_type, docname, ignore_permissions=True)
+
+	if restricted_docnames:
+		raise frappe.PermissionError("You do not have permisssion to delete " + ", ".join(restricted_docnames))
